Re-use owner as we intended, seems to have a positive effect when we use a good random selection, so use that and reduce owner to a short

+++ src/cell.h
@@ -461,7 +461,7 @@ struct cell {
   float dmin;
 
   /*! ID of the previous owner, e.g. runner. */
-  int owner;
+  short owner;
 
   /*! ID of a threadpool thread that maybe associated with this cell. */
   int tpid;
